Split code for the parser and add the mechanism for controlling the color.
